make powdr generic over the pgo method
