+++ autoprecompiles/src/adapter.rs
@@ -6,12 +6,48 @@ use powdr_number::FieldElement;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    blocks::{BasicBlock, Candidate, Instruction, Program},
+    blocks::{BasicBlock, Instruction, Program},
     constraint_optimizer::IsBusStateful,
     memory_optimizer::MemoryBusInteraction,
     range_constraint_optimizer::RangeConstraintHandler,
-    Apc, InstructionHandler, VmConfig,
+    Apc, InstructionHandler, PowdrConfig, VmConfig,
 };
+
+pub struct AdapterApcWithStats<A: Adapter> {
+    apc: AdapterApc<A>,
+    stats: Option<ApcStats<A>>,
+}
+impl<A: Adapter> AdapterApcWithStats<A> {
+    pub fn with_stats(mut self, stats: ApcStats<A>) -> Self {
+        self.stats = Some(stats);
+        self
+    }
+
+    pub fn into_parts(self) -> (AdapterApc<A>, Option<ApcStats<A>>) {
+        (self.apc, self.stats)
+    }
+}
+
+impl<A: Adapter> From<AdapterApc<A>> for AdapterApcWithStats<A> {
+    fn from(apc: AdapterApc<A>) -> Self {
+        Self { apc, stats: None }
+    }
+}
+
+pub trait PgoAdapter {
+    type Adapter: Adapter;
+
+    fn create_apcs_with_pgo(
+        &self,
+        blocks: Vec<BasicBlock<<Self::Adapter as Adapter>::Instruction>>,
+        config: &PowdrConfig,
+        vm_config: AdapterVmConfig<Self::Adapter>,
+    ) -> Vec<AdapterApcWithStats<Self::Adapter>>;
+
+    fn pc_execution_count(&self, _pc: u64) -> Option<u32> {
+        None
+    }
+}
 
 pub trait Adapter: Sized {
     type Field: Serialize + for<'de> Deserialize<'de> + Send + Clone;
@@ -22,7 +58,6 @@ pub trait Adapter: Sized {
         + IsBusStateful<Self::PowdrField>
         + RangeConstraintHandler<Self::PowdrField>
         + Sync;
-    type Candidate: Candidate<Self> + Send;
     type Program: Program<Self::Instruction> + Send;
     type Instruction: Instruction<Self::Field> + Serialize + for<'de> Deserialize<'de> + Send;
     type MemoryBusInteraction<V: Ord + Clone + Eq + Display + Hash>: MemoryBusInteraction<
@@ -30,6 +65,7 @@ pub trait Adapter: Sized {
         V,
     >;
     type CustomBusTypes: Clone + Display + Sync + Eq + PartialEq;
+    type ApcStats: Send + Sync;
 
     fn into_field(e: Self::PowdrField) -> Self::Field;
 
@@ -40,7 +76,7 @@ pub trait Adapter: Sized {
     }
 }
 
-pub type ApcStats<A> = <<A as Adapter>::Candidate as Candidate<A>>::ApcStats;
+pub type ApcStats<A> = <A as Adapter>::ApcStats;
 pub type AdapterApc<A> = Apc<<A as Adapter>::Field, <A as Adapter>::Instruction>;
 pub type AdapterVmConfig<'a, A> = VmConfig<
     'a,
