+++ openvm/src/lib.rs
@@ -29,6 +29,7 @@ use openvm_stark_sdk::engine::StarkFriEngine;
 use openvm_stark_sdk::openvm_stark_backend::p3_field::PrimeField32;
 use openvm_stark_sdk::p3_baby_bear::BabyBear;
 use powdr_autoprecompiles::evaluation::AirStats;
+use powdr_autoprecompiles::pgo::{CellPgo, InstructionPgo, NonePgo, PgoConfig};
 use powdr_autoprecompiles::{execution_profile::execution_profile, PowdrConfig};
 use powdr_extension::{PowdrExecutor, PowdrExtension, PowdrPeriphery};
 use powdr_openvm_hints_circuit::{HintsExecutor, HintsExtension, HintsPeriphery};
@@ -45,6 +46,7 @@ use std::{
     sync::Arc,
 };
 
+use crate::customize_exe::OpenVmApcCandidate;
 pub use crate::customize_exe::Prog;
 use tracing::Level;
 
@@ -62,7 +64,6 @@ pub mod symbolic_instruction_builder;
 mod utils;
 pub use opcode::instruction_allowlist;
 pub use powdr_autoprecompiles::DegreeBound;
-pub use powdr_autoprecompiles::PgoConfig;
 
 type BabyBearSC = BabyBearPoseidon2Config;
 
@@ -375,14 +376,48 @@ pub fn compile_exe_with_elf(
     pgo_config: PgoConfig,
 ) -> Result<CompiledProgram, Box<dyn std::error::Error>> {
     let elf = powdr_riscv_elf::load_elf_from_buffer(elf);
-    let compiled = customize(
-        original_program,
-        elf.text_labels(),
-        elf.debug_info(),
-        config,
-        implementation,
-        pgo_config,
-    );
+    let compiled = match pgo_config {
+        PgoConfig::Cell(pgo_data, max_total_columns) => {
+            let max_total_apc_columns: Option<usize> = max_total_columns.map(|max_total_columns| {
+                let original_config = OriginalVmConfig::new(original_program.vm_config.clone());
+
+                let total_non_apc_columns = original_config
+                    .chip_inventory_air_metrics(config.degree_bound.identities)
+                    .values()
+                    .map(|m| m.total_width())
+                    .sum::<usize>();
+                max_total_columns - total_non_apc_columns
+            });
+
+            customize(
+                original_program,
+                elf.text_labels(),
+                elf.debug_info(),
+                config,
+                implementation,
+                CellPgo::<_, OpenVmApcCandidate<_, _>>::with_pgo_data_and_max_columns(
+                    pgo_data,
+                    max_total_apc_columns,
+                ),
+            )
+        }
+        PgoConfig::Instruction(pgo_data) => customize(
+            original_program,
+            elf.text_labels(),
+            elf.debug_info(),
+            config,
+            implementation,
+            InstructionPgo::with_pgo_data(pgo_data),
+        ),
+        PgoConfig::None => customize(
+            original_program,
+            elf.text_labels(),
+            elf.debug_info(),
+            config,
+            implementation,
+            NonePgo::default(),
+        ),
+    };
     // Export the compiled program to a PIL file for debugging purposes.
     export_pil(
         &mut BufWriter::new(File::create("debug.pil").unwrap()),
