+++ openvm/tests/apc_builder.rs
@@ -1,8 +1,9 @@
 use openvm_instructions::instruction::Instruction;
 use openvm_sdk::config::SdkVmConfig;
 use openvm_stark_sdk::p3_baby_bear::BabyBear;
+use powdr_autoprecompiles::blocks::BasicBlock;
 use powdr_autoprecompiles::evaluation::evaluate_apc;
-use powdr_autoprecompiles::{build, BasicBlock, VmConfig};
+use powdr_autoprecompiles::{build, VmConfig};
 use powdr_number::BabyBearField;
 use powdr_openvm::bus_interaction_handler::OpenVmBusInteractionHandler;
 use powdr_openvm::extraction_utils::OriginalVmConfig;
